+++ changelog.md
@@ -1,6 +1,7 @@
 # ACEmulator Change Log
 
-### 2017-08-01
+
+### 2017-08-06
 [Og II]
 
 * Start inventory save.   
@@ -19,6 +20,16 @@
 * 3 implement view contents 1 per each container we have
 * 4 Add logic to drop item to manage shard data clean up - maybe via landblock.
 
+### 2017-08-02
+[Ripley]
+* NOTE: The following changes require ACE-World database v0.1.8 or newer...
+* Added NpcLooksLikeObject check for Creature assessment profile fix provided by @OptimShi.
+* Moved Default do nothing UseDone to WorldObject and removed it from GenericObject.
+* Added Cow weenie object.
+* Added Cow, Creature, and Container to the WorldObjectFactory.
+* Say Hi to the cow in Holtburg.
+* Load attributes stored in world database into objects.
+
 ### 2017-07-31
 [OptimShi]
 * Added check to ensure a player can't spend more attribute credits than they should at character creation.
@@ -33,8 +44,13 @@
 * You can also unlearn them and they will be removed.
 * Now, drag your spells into any configuration in your spell bars you like.   You can add them, remove them or reorder them.
 * They will be saved on our periodic save, or when you log out or you can force a save with the debug command @save-now.
+<<<<<<< HEAD
 * Added table to persist spell bar configuration.
 * That's all for this PR.   
+=======
+* Added table to persist spellbar configuration.
+* That's all for this PR.
+>>>>>>> 65fbd37a355f8014deaf8437bb5a1f6c354a27fe
 
 ### 2017-07-29
 [Ripley]